Flip asteroid position logic; enable saucer and fixes
The asteroid position logic in Asteroid.cs was flipped by changing
the condition to check if `rnd.Next(0, 2)` equals `1` instead of `0`.

In GameForm.cs, the particle effect `count` was increased from 15
to 100, and previously commented-out code for creating a `Saucer`
and initializing 5 asteroids was uncommented.

In Particle.cs, a null-conditional operator was added to the
`p.Draw(g)` call in the `DrawAll` method to prevent potential
`NullReferenceException` errors.